% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_functions.R
\name{calc_sigma_scaling}
\alias{calc_sigma_scaling}
\title{Sigma Scaling Plotter}
\usage{
calc_sigma_scaling(chain, plot = F)
}
\arguments{
\item{chain}{Vector of n length, where n is the number of trials or sampler iterations}

\item{plot}{Boolean. Whether to return a plot or the elements used to make it.}
}
\value{
If plot is true, a sigma scaling plot. If false, a vector with the standard deviations at each lag
}
\description{
Plots a scaling of the sd in the distribution of price changes across time lags and returns the value of the slope
}
\details{
Markets show sigma scaling exponents around 0.5.
}
\examples{

set.seed(1)
chain1 <- sampler_mh(1, "norm", c(0,1), diag(1))
calc_sigma_scaling(chain1[[1]])

set.seed(1)
chain1 <- sampler_mh(1, "norm", c(0,1), diag(1))
calc_sigma_scaling(chain1[[1]], plot = FALSE)
}
