% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotters.R
\name{change_1d}
\alias{change_1d}
\title{Change Calculator}
\usage{
change_1d(X)
}
\arguments{
\item{X}{one-dimensional chain.}
}
\value{
one-dimensional chain of step sizes
}
\description{
Given a one-dimensional chain, it returns a 1d chain of step sizes and direction (e.g. given a chain \[1,4,1\], it returns \[3, -3\])
}
\examples{
chain <- stats::runif(10, 1, 10)
change_1d(chain)


}
