% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABSClass.R
\name{CoreABS}
\alias{CoreABS}
\title{CoreABS Object}
\description{
This is the parent \link[R6:R6Class]{R6} class of the Auto-correlated Bayesian Sampler \insertCite{@ABS, @zhuAutocorrelatedBayesian2023}{samplr}. It is a sequential sampling model assuming people draw autocorrelated samples from memory or beliefs, i.e., posterior of hypotheses.
}
\references{
\insertRef{zhuAutocorrelatedBayesian2023}{samplr}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{lambda}}{The rate parameter of the gamma distribution for decision time.}

\item{\code{n_chains}}{The number of chains of the sampler. It should be an integer.}

\item{\code{distr_name}}{The type of the posterior hypothesis distribution.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CoreABS-new}{\code{CoreABS$new()}}
\item \href{#method-CoreABS-clone}{\code{CoreABS$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoreABS-new"></a>}}
\if{latex}{\out{\hypertarget{method-CoreABS-new}{}}}
\subsection{Method \code{new()}}{
Create a new 'CoreABS' object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoreABS$new(lambda, n_chains, distr_name = "norm")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lambda}}{The rate parameter of the gamma distribution for decision time.}

\item{\code{n_chains}}{The number of chains of the sampler. It should be an integer.}

\item{\code{distr_name}}{The type of the posterior hypothesis distribution.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new 'CoreABS' object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoreABS-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CoreABS-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoreABS$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
