% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r3a.R
\name{d2dunif}
\alias{d2dunif}
\title{d2dunif}
\usage{
d2dunif(x, y, min = 0, max = 1)
}
\arguments{
\item{x, y}{vector of quantiles.}

\item{min, max}{lower and upper limits of the distribution on either side. Must be finite.}
}
\value{
a numeric density
}
\description{
This function returns the density for a 2D uniform distribution defined on a square.
}
\examples{
d2dunif(x = 0.5, y = 0.5)
d2dunif(x = 0, y = 2, min = 0, max = 2)
}
