% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotters.R
\name{euc_d}
\alias{euc_d}
\title{Euclidean Distance Calculator}
\usage{
euc_d(x, y)
}
\arguments{
\item{x, y}{Vector or Matrix. If vector of length d, a d-dimensional point. If matrix of size n x d, n points of d dimensions.}
}
\value{
Numeric or Vector. Distance between points. If vector, distances between the ith points in the matrices.
}
\description{
Calculates the distance between two points. If instead given two matrices A and B of size n x d, for n number of items of d dimensions each, it will return an n - 1 length vector with the distances between each of the items (i.e. the distance between the item A[1,] and B[1,]; between A[2,] and B[2,], and so on).
}
\details{
It assumes that all dimensions have the same units or that all dimensions are normalized (mean = 0 and std = 1). Adapted from \href{https://hlab.stanford.edu/brian/euclidean_distance_in.html}{here}
}
\examples{
# distance between two points
euc_d(c(0,0), c(3,4))

# ask for multiple distances
M <- matrix(0, nrow = 3, ncol = 2)
M2 <- matrix(0, nrow = 3, ncol = 2)
for (i in 1:3){
  M[i,] <- stats::runif(2) * 5
  M2[i,] <- stats::runif(2) * 5
}
euc_d(M, M2)
}
