% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectq3c.R
\name{find_quantile_2d}
\alias{find_quantile_2d}
\title{find_quantile_2d}
\usage{
find_quantile_2d(jpdf, ...)
}
\arguments{
\item{jpdf}{a function that is the joint pdf of the random variables.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
a quantile in the support of the distribution
}
\description{
This function trials quantiles until one is found with positive density.
}
