% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectq3b.R
\name{find_support_element}
\alias{find_support_element}
\title{find_support_element}
\usage{
find_support_element(pdf, ...)
}
\arguments{
\item{pdf}{a function that is the pdf of the random variable.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
a quantile in the support of the distribution
}
\description{
find_support_element
}
