% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectq3a.R
\name{jdbeta}
\alias{jdbeta}
\title{jdbeta}
\usage{
jdbeta(x, y, shape1, shape2)
}
\arguments{
\item{x, y}{vector of quantiles}

\item{shape1, shape2}{non-negative parameters of the Beta distribution.}
}
\value{
numeric vector of joint densities
}
\description{
jdbeta
}
\examples{
jdbeta(x = 0.5, y = 0.5, shape1 = 2, shape2 = 2)
jdbeta(x = 0.2, y = 0.2, shape1 = 2, shape2 = 5)
jdbeta(x = runif(3), y = runif(3), shape1 = 2, shape2 = 2)
}
