% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectq3a.R
\name{jdcirclecontour}
\alias{jdcirclecontour}
\title{jdcirclecontour}
\usage{
jdcirclecontour(x, y)
}
\arguments{
\item{x, y}{vector of quantiles}
}
\value{
numeric vector of joint densities
}
\description{
This function returns the joint density = (3/8)(x^2 + y^2) of two random variables defined on a square from -1 to 1.
}
\examples{
jdcirclecontour(x = 0, y = 0)
jdcirclecontour(x = 1, y = -1)
jdcirclecontour(x = runif(n = 3, min = -1, max = 1),
                y = runif(n = 3, min = -1, max = 1))
}
