% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectq3c.R
\name{jdnorm}
\alias{jdnorm}
\title{jdnorm}
\usage{
jdnorm(x, y, mean = c(0, 0), sd = c(1, 1), rho = 0)
}
\arguments{
\item{x, y}{vector of quantiles}

\item{mean}{vector of both means}

\item{sd}{vector of both standard deviations}

\item{rho}{correlation between the two normals}
}
\value{
numeric vector of joint densities
}
\description{
This function returns the joint density from a bivariate random normal.
}
\examples{
## center of standard bivariate normal
jdnorm(x = 0, y = 0)

## change scale
jdnorm(x = 0, y = 0, sd = c(10, 10))

## change location
jdnorm(x = 5, y = 5, mean = c(5, 5))

## scale-location transform
jdnorm(x = 42, y = 42, mean = c(42, 42), sd = c(10, 10))

## multiple samples from correlated
jdnorm(x = c(1, -1, -1, 1), y = c(1, 1, -1, -1), rho = 0.5)
}
