% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectq3a.R
\name{jdunif}
\alias{jdunif}
\title{jdunif}
\usage{
jdunif(x, y, min = 0, max = 1)
}
\arguments{
\item{x, y}{vector of quantiles}

\item{min, max}{lower and upper limits of the distribution for each side. Must be finite.}
}
\value{
numeric vector of joint densities
}
\description{
jdunif
}
\examples{
jdunif(x = 0.5, y = 0.5)
jdunif(x = 5, y = 5, min = 0, max = 10)
jdunif(x = runif(3), y = runif(3))
}
