% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABSClass.R
\name{LiABS}
\alias{LiABS}
\title{Auto-correlated Bayesian Sampler by Li}
\description{
This Auto-correlated Bayesian Sampler (ABS) model is developed by Yun-Xiao Li following the ABS of Zhu. LiABS has an extra component of non-decision time when calculating the response time.
}
\examples{

## ------------------------------------------------
## Method `LiABS$two_alt_force_choice`
## ------------------------------------------------

abs_model <- LiABS$new(delta=5, lambda=20, n_chains=3, width=1, nd_time = 0.1, s_nd_time = 0.3)
tafc_sim <- abs_model$two_alt_force_choice(dec_bdry=0, discrim=1, trial_fdbk=c(0, 0, 1, 0, 1))

}
\section{Super class}{
\code{\link[samplr:CoreABS]{samplr::CoreABS}} -> \code{LiABS}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{nd_time}}{The non-decision time.}

\item{\code{s_nd_time}}{The range of the non-decision time.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LiABS-new}{\code{LiABS$new()}}
\item \href{#method-LiABS-two_alt_force_choice}{\code{LiABS$two_alt_force_choice()}}
\item \href{#method-LiABS-clone}{\code{LiABS$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="samplr" data-topic="CoreABS" data-id="confidence"><a href='../../samplr/html/CoreABS.html#method-CoreABS-confidence'><code>samplr::CoreABS$confidence()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LiABS-new"></a>}}
\if{latex}{\out{\hypertarget{method-LiABS-new}{}}}
\subsection{Method \code{new()}}{
Create a new 'LiABS' object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LiABS$new(
  nd_time,
  s_nd_time,
  delta,
  lambda,
  n_chains,
  width,
  distr_name = "norm",
  mc3_iterations = 100
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{nd_time}}{The non-decision time.}

\item{\code{s_nd_time}}{The range of the non-decision time.}

\item{\code{delta}}{The threshold of the relative difference. It should be an integer.}

\item{\code{lambda}}{The rate parameter of the gamma distribution for decision time.}

\item{\code{n_chains}}{The number of chains of the sampler. It should be an integer.}

\item{\code{width}}{The proposal width of the sampler.}

\item{\code{distr_name}}{The type of the posterior hypothesis distribution.}

\item{\code{mc3_iterations}}{The number of iterations for each chunk. See details for more information.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new 'LiABS' object.

ABS simulating two-alternative-force choice task
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LiABS-two_alt_force_choice"></a>}}
\if{latex}{\out{\hypertarget{method-LiABS-two_alt_force_choice}{}}}
\subsection{Method \code{two_alt_force_choice()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LiABS$two_alt_force_choice(dec_bdry, discrim, trial_fdbk)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dec_bdry}}{The decision boundary that separates the posterior hypothesis distribution.}

\item{\code{discrim}}{The stimuli discriminability.}

\item{\code{trial_fdbk}}{The feedback of each trial. It should be a numeric vector only consist of 0 (below the decision boundary) and/or 1 (above the decision boundary).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data frame with ten columns:
\enumerate{
\item{trial: the index of trials;}
\item{samples: the samples of ABS sampler for the trial;}
\item{support: the relative difference between the numbers of supporting samples for each response;}
\item{length: the length of the sequence;}
\item{response: the response predicted by ABS. 0 represents the response coded as below the decision boundary and 1 represent the opposite;}
\item{feedback: the feedback or the true stimuli of the experiment. 0 represents the response coded as below the decision boundary and 1 represent the opposite;}
\item{accuracy: whether the response is the same as the feedback. 0 represents error, and 1 represents correct;}
\item{nd_time: the non-decision time for this trial;}
\item{rt: the response time, including both the non-decision and the decision time;}
\item{rept: whether the next trial repeats the same response as the current trial;}
}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{abs_model <- LiABS$new(delta=5, lambda=20, n_chains=3, width=1, nd_time = 0.1, s_nd_time = 0.3)
tafc_sim <- abs_model$two_alt_force_choice(dec_bdry=0, discrim=1, trial_fdbk=c(0, 0, 1, 0, 1))

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LiABS-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LiABS-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LiABS$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
