% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotters.R
\name{plot_all}
\alias{plot_all}
\title{Plotter Wrapper}
\usage{
plot_all(chain, title = NULL)
}
\arguments{
\item{chain}{Vector of n length, where n is the number of trials or sampler iterations}

\item{title}{Title of the uberplot}
}
\value{
A grid plotting all the plot_* functions
}
\description{
Plots all the plot_* plots into a grid for ease.
}
\examples{
set.seed(1)
chain1 <- sampler_mh(1, "norm", c(0,1), diag(1))
plot_all(chain1[[1]])
}
