% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotters.R
\name{plot_autocorr}
\alias{plot_autocorr}
\title{Autocorrelation Plotter}
\usage{
plot_autocorr(chain, changeACF = TRUE, alpha = 0.05, lag.max = 100)
}
\arguments{
\item{chain}{Vector of n length, where n is the number of trials or sampler iterations}

\item{changeACF}{Boolean. If true, plot the autocorrelation of the change series. If false, plot the autocorrelation of the given chain.}

\item{alpha}{Measure of Type I error - defaults to .05}

\item{lag.max}{Length of the x axis. How far to examine the lags.}
}
\value{
An autocorrelation plot
}
\description{
Plots the autocorrelation of a given sequence, or of the size of the steps (returns).
}
\details{
Markets display no significant autocorrelations in the returns of a given asset.
}
\examples{
target <- distr::Norm()
pd_func <- make_distr_pdf(target)
chain1 <- sampler_mcmc(pd_func, 0, sigma_prop = 1)
plot_autocorr(chain1[[1]])
}
