% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotters.R
\name{plot_change}
\alias{plot_change}
\title{Change Plotter}
\usage{
plot_change(chain)
}
\arguments{
\item{chain}{Vector of n length, where n is the number of trials or sampler iterations}
}
\value{

}
\description{
Plots a change series against iterations. Useful to see if there is clustering of volatility in returns, like occurs in financial markets
}
\examples{
target <- distr::Norm()
pd_func <- make_distr_pdf(target)
chain1 <- sampler_mcmc(pd_func, 0, sigma_prop = 1)
plot_change(chain1[[1]])
}
