% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotters.R
\name{plot_levy}
\alias{plot_levy}
\title{Levy Flights Plotter}
\usage{
plot_levy(chain, plot = TRUE)
}
\arguments{
\item{plot}{Boolean. plot Boolean. Whether to return a plot or the elements used to make it.}
}
\value{
If plot is true, it returns a simple plot with the log absolute difference in estimates and their frequency, as well as an estimate for the \eqn{\mu} parameter. If false it returns a list with what's required to make the plot.
}
\description{
This plotter analyses if the length of the jumps the sampler is making (\eqn{l}) belongs to a Levy probability density distribution, \eqn{P(l) \approx l^{-\mu}}.
}
\details{
Values of \eqn{\mu \approx 2} have been used to describe foraging in animals, and produce the most effective foraging [(Viswanathan et al., 1999)](https://www.nature.com/articles/44831). See [Zhu et al. 2018](https://dl.acm.org/doi/abs/10.5555/3327345.3327477) for a comparison of Levy Flight and PSD measures for different samplers in multimodal representations.

 @param chain Matrix of n x d dimensions, n = iterations, d = dimensions
}
\examples{
target <- distr::Norm()
pd_func <- make_distr_pdf(target)
chain1 <- sampler_mcmc(pd_func, 0, sigma_prop = 1)
plot_levy(chain1[[1]])
}
