% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotters.R
\name{plot_qqplot}
\alias{plot_qqplot}
\title{QQ-Plotter}
\usage{
plot_qqplot(chain, change = TRUE)
}
\arguments{
\item{chain}{Vector of n length, where n is the number of trials or sampler iterations}

\item{change}{Boolean. If false, it plots a qqplot of the given chain. If true, it creates a chain of step sizes (using \link[SampleR]{change_1d})}
}
\value{
QQ plot of Theoretical vs Empirical values
}
\description{
Plots a QQ plot of Empirical values against Theoretical values from a normal distribution. Can plot the chain points or the distances between successive points
}
\examples{
set.seed(1)
chain1 <- sampler_mh(1, "norm", c(0,1), diag(1))
plot_qqplot(chain1[[1]])
}
