% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotters.R
\name{plot_series}
\alias{plot_series}
\title{Series Plotter}
\usage{
plot_series(chain)
}
\arguments{
\item{chain}{Vector of n length, where n is the number of trials or sampler iterations}
}
\value{
A series plot
}
\description{
Plots the value of a one-dimensional series against the iteration where it occurred. Useful to see the general pattern of the chain (white noise, random walk, volatility clustering)
}
\examples{
target <- distr::Norm()
pd_func <- make_distr_pdf(target)
chain1 <- sampler_mcmc(pd_func, 0, sigma_prop = 1)
plot_series(chain1[[1]])
}
