% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotters.R
\name{plot_sigma_scaling}
\alias{plot_sigma_scaling}
\title{Sigma Scaling Plotter}
\usage{
plot_sigma_scaling(chain, plot = TRUE)
}
\arguments{
\item{chain}{Vector of n length, where n is the number of trials or sampler iterations}

\item{plot}{Boolean. Whether to return a plot or the elements used to make it.}
}
\value{
If plot is true, a sigma scaling plot. If false, a vector with the standard deviations at each lag
}
\description{
Plots a scaling of the sd in the distribution of price changes across time lags and returns the value of the slope
}
\details{
Markets show sigma scaling exponents around 0.5.
}
\examples{

target <- distr::Norm()
pd_func <- make_distr_pdf(target)
chain1 <- sampler_mcmc(pd_func, 0, sigma_prop = 1)
plot_sigma_scaling(chain1[[1]])

set.seed(1)
target <- distr::Norm()
pd_func <- make_distr_pdf(target)
chain1 <- sampler_mcmc(pd_func, 0, sigma_prop = 1)
plot_sigma_scaling(chain1[[1]], plot = FALSE)
}
