% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectq2a.R
\name{projectq2a}
\alias{projectq2a}
\title{projectq2a}
\usage{
projectq2a(n, pdf, a, b, C, ...)
}
\arguments{
\item{n}{number of observations. If length(n) > 1, the length is taken to be the number required.}

\item{pdf}{a function that is the pdf of the random variable.}

\item{a, b}{lower and upper limits of the distribution. Must be finite such that \eqn{P(a \le X \le b) = 1}.}

\item{C}{a numeric such that \eqn{f(x) \le C} for all values of x.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
numeric vector of random deviates
}
\description{
This function generates random deviates from a continuous random variable with the supplied probability density function via rejection sampling.
}
\examples{
## sample from standard uniform
projectq2a(n = 1, pdf = dunif, a = 0, b = 1, C = 1, min = 0, max = 1)

## plot many samples from standard uniform as densities
hist(projectq2a(n = 10000, pdf = dunif, a = 0, b = 1, C = 1, min = 0, max = 1), probability = TRUE)
curve(dunif, col = "red", add = TRUE)

## sample from beta(2, 2)
projectq2a(n = 1, pdf = dbeta, a = 0, b = 1, C = 1.5, shape1 = 2, shape2 = 2)

## plot many samples from beta(2, 2)
hist(projectq2a(n = 10000, pdf = dbeta, a = 0, b = 1, C = 1.5, shape1 = 2, shape2 = 2), probability = TRUE)
curve(dbeta(x, shape1 = 2, shape2 = 2), col = "red", add = TRUE)
}
