% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectq3b.R
\name{projectq3b}
\alias{projectq3b}
\title{projectq3b}
\usage{
projectq3b(n, pdf, a, b, ...)
}
\arguments{
\item{n}{number of observations. If length(n) > 1, the length is taken to be the number required.}

\item{pdf}{a function that is the pdf of the random variable.}

\item{a, b}{lower and upper limits of the distribution. Must be finite such that \eqn{P(a \le X \le b) = 1}.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
numeric vector of semi-random deviates
}
\description{
This function generates semi-random deviates from a continuous random variable with the supplied probability density function via Markov Chain Monte Carlo.
}
\examples{
## sample from standard uniform
deviates <- projectq3b(n = 10000, pdf = dunif, a = 0, b = 1)
hist(deviates, probability = TRUE)
curve(dunif(x), col = "red", add = TRUE)

## sample from beta(2, 2)
deviates <- projectq3b(n = 10000, pdf = dbeta, a = 0, b = 1, shape1 = 2, shape2 = 2)
hist(deviates, probability = TRUE)
curve(dbeta(x = x, shape1 = 2, shape2 = 2), col = "red", add = TRUE)

## sample from custom random variable
dcustom <- function(x) x / 2
deviates <- projectq3b(n = 10000, pdf = dcustom, a = 0, b = 2)
hist(deviates, probability = TRUE)
curve(dcustom, add = TRUE)
}
