% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectq3c.R
\name{projectq3c}
\alias{projectq3c}
\title{projectq3c}
\usage{
projectq3c(n, jpdf, ...)
}
\arguments{
\item{n}{number of observations. If length(n) > 1, the length is taken to be the number required.}

\item{jpdf}{a function that is the joint pdf of the distribution}

\item{...}{further arguments passed to or from other methods.}
}
\value{
data.frame of semi-random deviates
}
\description{
This function generates semi-random deviates from a continuous bivariate distribution with the supplied joint probability density function via Markov Chain Monte Carlo.
}
\examples{
plot(projectq3c(n = 10000, jpdf = jdunif))
plot(projectq3c(n = 10000, jpdf = jdunif, min = -1, max = 1))
plot(projectq3c(n = 10000, jpdf = jdcirclecontour))
plot(projectq3c(n = 10000, jpdf = jdnorm, r = 0.75))
}
