% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r3a.R
\name{r3a}
\alias{r3a}
\title{r3a}
\usage{
r3a(n, pdf, a, b, C, ...)
}
\arguments{
\item{n}{number of observations. If length(n) > 1, the length is taken to be the number required.}

\item{pdf}{a function that is the joint pdf of the distribution. Should use x and y as quantile arguments.}

\item{a, b}{lower and upper limits of the sides of the distribution. Must be finite such that \eqn{P(a \le X \le b \cap a \le Y \le b) = 1}.}

\item{C}{a numeric such that \eqn{f(x,y) \le C} for all values of x and y.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
data.frame of random deviate pairs
}
\description{
This function generates pairs of random deviates from a 2D continuous distribution defined on a square with the supplied joint probability density function via rejection sampling.
}
\examples{
plot(r3a(n = 10000, pdf = d2dunif, a = 0, b = 1, C = 1))
plot(r3a(n = 10000, pdf = d2dunif, a = 0, b = 2, C = 1/4, min = 0, max = 2))
}
