% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rABS.R
\name{rABS}
\alias{rABS}
\title{ABS simulation}
\usage{
rABS(
  dec_bdry,
  discrim,
  delta,
  nd_time,
  s_nd_time,
  lambda,
  trial_fdbk,
  nChains,
  width,
  distr_name = "norm",
  mc3_iterations = 100
)
}
\arguments{
\item{dec_bdry}{The decision boundary that separates the posterior hypothesis distribution.}

\item{discrim}{The stimuli discriminability.}

\item{delta}{The threshold of the relative difference. It should be an integer.}

\item{nd_time}{The non-decision time.}

\item{s_nd_time}{The range of the non-decision time.}

\item{lambda}{The rate parameter of the gamma distribution for decision time.}

\item{trial_fdbk}{The feedback of each trial. It should be a numeric vector only consist of 0 (below the decision boundary) and/or 1 (above the decision boundary).}

\item{nChains}{The number of chains of the sampler. It should be an integer.}

\item{width}{The proposal width of the sampler.}

\item{distr_name}{The type of the posterior hypothesis distribution.}

\item{mc3_iterations}{The number of iterations for each chunk. See details for more information.}
}
\value{
A data frame with ten columns:
\enumerate{
\item{trial: the index of trials;}
\item{samples: the samples of ABS sampler for the trial;}
\item{support: the relative difference between the numbers of supporting samples for each response;}
\item{length: the length of the sequence;}
\item{response: the response predicted by ABS. 0 represents the response coded as below the decision boundary and 1 represent the opposite;}
\item{feedback: the feedback or the true stimuli of the experiment. 0 represents the response coded as below the decision boundary and 1 represent the opposite;}
\item{accuracy: whether the response is the same as the feedback. 0 represents error, and 1 represents correct;}
\item{nd_time: the non-decision time for this trial;}
\item{rt: the response time, including both the non-decision and the decision time;}
\item{rept: whether the next trial repeats the same response as the current trial;}
}
}
\description{
This function is for simulating the Auto-correlated Bayesian Sampler \insertCite{@ABS, @zhuAutocorrelatedBayesian2023}{samplr}. It is a sequential sampling model assuming people draw autocorrelated samples from memory or beliefs, i.e., posterior of hypotheses.
}
\details{
The \code{rABS} function in R runs simulations in chunks of fixed length determined by the \code{mc3_iterations} argument. In each trial simulation, the function generates the first chunk and checks if the simulated sequence meets the stopping rule within that chunk. If the stopping rule is met, the function cuts the chunk at that point, returns the sequence, and starts a new trial simulation. If the stopping rule is not met, the function proceeds to the next chunk, beginning where the previous chunk ended, and repeats the entire process.
}
\examples{
simulation <- rABS(dec_bdry=0, discrim=1, delta=2, nd_time=0.3, s_nd_time=0.5, lambda=6, trial_fdbk=c(0, 0, 1, 0, 1), nChains=3, width=1)

}
\references{
\insertRef{zhuAutocorrelatedBayesian2023}{samplr}
}
