% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_q2ab.R
\name{rdist}
\alias{rdist}
\title{rdist}
\usage{
rdist(n, pdf, a, b, C, ...)
}
\arguments{
\item{n}{number of observations. If length(n) > 1, the length is taken to be the number required.}

\item{pdf}{a function that is the pdf of the random variable.}

\item{a, b}{lower and upper limits of the distribution. Must be finite such that $P(a \le X \le b) = 1$.}

\item{C}{a numeric such that $f(x) \le C$ for all values of x.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
numeric vector of random deviates
}
\description{
This function generates random deviates from a continuous random variable with the supplied probability density function via rejection sampling.
}
\examples{
hist(rdist(n = 1000000, pdf = dunif, a = 0, b = 1, C = 1, min = 0, max = 1), probability = TRUE)
}
