% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sample_to_indicator}
\alias{sample_to_indicator}
\title{Transform a sample vector into an inclusion indicator vector}
\usage{
sample_to_indicator(sample, population_size)
}
\arguments{
\item{sample}{A vector of sample indices.}

\item{population_size}{The total size of the population.}
}
\description{
Transform a sample vector into an inclusion indicator vector
}
