% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplers.R
\name{sampler_hmc}
\alias{sampler_hmc}
\title{Hamiltonian Monte-Carlo Sampler.}
\usage{
sampler_hmc(
  distr_name,
  distr_params,
  start,
  epsilon = 0.5,
  L = 10,
  iterations = 1024,
  weights = NULL
)
}
\arguments{
\item{distr_name}{Name of the distribution from which to sample from.}

\item{distr_params}{Distribution parameters.}

\item{start}{Vector. Starting position of the sampler.}

\item{epsilon}{Size of the leapfrog step}

\item{L}{Number of leapfrog steps per iteration}

\item{iterations}{Number of iterations of the sampler.}

\item{weights}{If using a mixture distribution, the weights given to each constituent distribution. If none given, it defaults to equal weights for all distributions.}
}
\description{
Hamiltonian Monte-Carlo, also called Hybrid Monte Carlo, is a sampling algorithm that uses Hamiltonian Dynamics to approximate a posterior distribution. Unlike MCMC and MC3, HMC uses not only the current position, but also a sense of momentum, to draw future samples. An introduction to HMC can be read \href{http://arxiv.org/abs/1701.02434}{here}
}
\details{
This implementations assumes that the momentum is drawn from a normal distribution with mean 0 and identity covariance matrix (p ~ N (0, I)). Hamiltonian Monte Carlo does not support discrete distributions.
}
\examples{

HMC <- sampler_hmc(distr_name = "norm", distr_params = c(0,1), start = 1, epsilon = .01, L = 100)
}
