% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplers.R
\name{sampler_mc3}
\alias{sampler_mc3}
\title{Metropolis-coupled MCMC sampler (MC3)}
\usage{
sampler_mc3(
  distr_name,
  distr_params,
  start,
  sigma_prop = NULL,
  nChains = 6,
  delta_T = 4,
  swap_all = TRUE,
  iterations = 1024L,
  weights = NULL
)
}
\arguments{
\item{distr_name}{Name of the distribution from which to sample from.}

\item{distr_params}{Distribution parameters.}

\item{start}{Vector. Starting position of the sampler.}

\item{sigma_prop}{Covariance matrix of the proposal distribution. If sampling in 1D space, it can be instead a number.}

\item{nChains}{Number of chains to run.}

\item{delta_T}{numeric, >1. Temperature increment parameter. The bigger this number, the steeper the increase in temperature between the cold chain and the next chain}

\item{swap_all}{Boolean. If true, every iteration attempts floor(nChains / 2) swaps. If false, only one swap per iteration.}

\item{iterations}{Number of iterations of the sampler.}

\item{weights}{If using a mixture distribution, the weights given to each constituent distribution. If none given, it defaults to equal weights for all distributions.}
}
\description{
This sampler is a variant of MCMC in which multiple parallel chains are run at different temperatures. The chains stochastically swap positions which allows the coldest chain to visit regions far from its starting point (unlike in MCMC). Because of this, an MC3 sampler can explore far-off regions, whereas an MCMC sampler may become stuck in a particular point of high density.
}
\examples{

# Sample from a normal distribution
mc_3 <- sampler_mc3(distr_name = "norm", distr_params = c(0,1), start = 1, sigma_prop = diag(1))
}
