% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplr-package.R
\docType{package}
\name{samplr-package}
\alias{samplr}
\alias{samplr-package}
\title{samplr: Compare Human Performance to Sampling Algorithms}
\description{
Understand human performance from the perspective of sampling, both looking at how people generate samples and how people use the samples they have generated. A longer overview and other resources can be found at \url{https://sampling.warwick.ac.uk}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://lucas-castillo.github.io/samplr/}
  \item Report bugs at \url{https://github.com/lucas-castillo/samplr/issues}
}

}
\author{
\strong{Maintainer}: Lucas Castillo \email{lucas.castillo@warwick.ac.uk} (\href{https://orcid.org/0000-0003-0274-0777}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Yun-Xiao Li \email{yunxiao.li@warwick.ac.uk} (\href{https://orcid.org/0000-0002-3509-6618}{ORCID}) [copyright holder]
  \item Adam N Sanborn \email{a.n.sanborn@warwick.ac.uk} (\href{https://orcid.org/0000-0003-0442-4372}{ORCID}) [copyright holder]
}

Other contributors:
\itemize{
  \item European Research Council (ERC) [funder]
}

}
\keyword{internal}
