% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sundh2023e3}
\alias{sundh2023e3}
\title{Data from Experiment 3 in \insertCite{sundh2023UnifiedExplanationVariability}{samplr}}
\format{
An object of class \code{data.frame} with 12420 rows and 10 columns.
}
\usage{
sundh2023e3
}
\description{
\describe{
\item{ID}{}
\item{block}{3 blocks in total}
\item{trial}{Trial Number within a block}
\item{query, querydetail}{Verbal descriptions of the query}
\item{querytype}{Type of query: e.g. notBgA = p(¬B|A)}
\item{Estimate}{Estimated probability, in percentages}
\item{starttime, endtime}{}
\item{RT}{}
}
}
\keyword{datasets}
