% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw7q2.R
\name{hw7q2}
\alias{hw7q2}
\alias{dhw7q2}
\alias{phw7q2}
\alias{qhw7q2}
\alias{rhw7q2}
\title{Homework 7 Question 2 Distribution}
\usage{
dhw7q2(x, log = FALSE)
phw7q2(q, lower.tail = TRUE, log.p = FALSE)
qhw7q2(p, lower.tail = TRUE, log.p = FALSE)
rhw7q2(n)

dhw7q2(x, log = FALSE)

phw7q2(q, lower.tail = TRUE, log.p = FALSE)

qhw7q2(p, lower.tail = TRUE, log.p = FALSE)

rhw7q2(n)
}
\arguments{
\item{lower.tail}{logical; if TRUE (default), probabilites are $P[X \le x]$ otherwise, $P[X > x]$}

\item{p}{vector of probabilites}

\item{n}{number of observations. If length(n) > 1, the length is taken to be the number required}

\item{[x, }{q] vector of quantiles}

\item{[log, }{log.p] logical; if TRUE, probabilities p are given as log(p)}
}
\value{
dnorm gives the density, pnorm gives the distribution function, qnorm gives the quantile function, and rnorm generates random deviates.
The length of the result is determined by n for rnorm, and is the maximum of the lengths of the numerical arguments for the other functions.
The numerical arguments other than n are recycled to the length of the result. Only the first elements of the logical arguments are used.
}
\description{
These functions provide information about the Homework 7
Question 2 distribution. dhw7q2 gives the density, phw7q2 gives the
distribution function, qhw7q2 gives the quantile function, and rhw7q2
generates random deviates.
}
\details{
This distribution has density $$f(x) = 6x(1-x)$$ for 0 < x < 1
}
\note{
The characteristics of output from pseudo-random number generators
(such as precision and periodicity) vary widely. See .Random.seed for more
information on R's random number generation algorithms.
}
\examples{
## density function
plot(dhw7q2(seq(0, 1, 0.0001)))

## distribution function
plot(phw7q2(seq(0, 1, 0.0001)))

## quantile function
qhw7q2(p = 0)
qhw7q2(p = 0, lower.tail = FALSE)
qhw7q2(p = seq(0, 1, 0.25))
plot(qhw7q2(p = seq(0, 1, 0.001)))

## random deviates
rhw7q2(n = 10)

}
