% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_functions.R
\name{plot_series}
\alias{plot_series}
\title{Series Plotter}
\usage{
plot_series(chain, change = FALSE)
}
\arguments{
\item{chain}{Vector of n length, where n is the number of trials or sampler iterations}

\item{change}{Boolean. Whether to plot the series of values or the series of changes between values.}
}
\value{
A series plot
}
\description{
Plots the value of a one-dimensional series against the iteration where it occurred. Useful to see the general pattern of the chain (white noise, random walk, volatility clustering)
}
\examples{
set.seed(1)
chain1 <- sampler_mh(1, "norm", c(0,1), diag(1))
plot_series(chain1[[1]])
}
