% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samp1D.R
\name{samp1D}
\alias{samp1D}
\title{Single Variable Rejection Sampler}
\usage{
samp1D(f, N, lb, ub, maxf)
}
\arguments{
\item{f}{the pdf that we are sampling from}

\item{N}{the number of attempted samples.}

\item{lb}{the lower bound of support f}

\item{ub}{upper bound of support of f}

\item{maxf}{bound on f}
}
\value{
A vector containg the samples form pdf
}
\description{
This function immplents single variable rejection sampling for rvs with bounded support and which have bounded pdf.
}
\details{
This paragrpah will show up somewhere else for additional information.
}
\examples{
betaPDF <- function(x) {
    ifelse(0 < x & x < 1, 2*x, 0)
}
hist(samp1D(betaPDF, 10000, 0, 1, 2))

}
