% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samp1D.R
\name{samp1D}
\alias{samp1D}
\title{Single Variable Rejection Sampler}
\usage{
samp1D(fun, a, b, N)
}
\arguments{
\item{fun}{the pdf that we are sampling from, input as a string. The variable sampled from must be input using the variable "x".}

\item{a}{The lower bound of the pdf.}

\item{b}{The upper bound of the pdf.}

\item{N}{The number of samples output by the sampling function.}
}
\value{
A vector containing the samples form pdf
}
\description{
This function immplents single variable rejection sampling for rvs with bounded support and which have bounded pdf.
}
\examples{

hist(samp1D("2*x", 0, 1, 10000))

}
