% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samp1D.R
\name{samp1D}
\alias{samp1D}
\title{Single Variable Rejection Sampler}
\usage{
samp1D(fun, a, b, lb, ub, N)
}
\arguments{
\item{fun}{the pdf that we are sampling from, input as an expression}

\item{a}{the lower bound on the function f}

\item{b}{the upper bound on the function f}

\item{lb}{the lower bound of support f}

\item{ub}{upper bound of support of f}

\item{N}{the number of attempted samples.}
}
\value{
A vector containg the samples form pdf
}
\description{
This function immplents single variable rejection sampling for rvs with bounded support and which have bounded pdf.
}
\details{
This paragrpah will show up somewhere else for additional information.
}
\examples{

hist(samp1D(expression(2*x), 0, 1, 0, 1, 10000))

}
