% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samp1D.R
\name{samp1D}
\alias{samp1D}
\title{Single Variable Rejection Sampler}
\usage{
samp1D(f, N)
}
\arguments{
\item{N}{The number of samples output by the sampling function.}

\item{fun}{The pdf that we are sampling from, input as a predefined function of \code{x}. For bounded support, the bounds must be input as a decimal.}
}
\value{
A vector containing the samples from the pdf.
}
\description{
This function implements single variable rejection sampling for rvs with bounded support and which have bounded pdf.
}
\examples{

f <- function(x) {
  ifelse(0 < x & x < 1, 2*x, 0)
}
hist(samp1D(f,10000))

f <- function(x) {
  ifelse(0 < x & x < 2, 1/2*x, 0)
}
hist(samp1D(f,10000))

f <- function(x) {
  ifelse(0 < x & x < 6.2832, 1/2/pi*(sin(x) + 1), 0)
}
hist(samp1D(f,10000))

f <- function(x) {
  ifelse(x < 0, 1/2*exp(x), 1/2*exp(-x))
}
hist(samp1D(f,10000))
}
