% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplers.R
\name{sampler_nuts}
\alias{sampler_nuts}
\title{No U-Turn Sampler.}
\usage{
sampler_nuts(
  distr_name,
  distr_params,
  start,
  epsilon = 0.5,
  delta_max = 1000,
  iterations = 1024
)
}
\arguments{
\item{distr_name}{Name of the distribution from which to sample from.}

\item{distr_params}{Distribution parameters.}

\item{start}{Vector. Starting position of the sampler.}

\item{epsilon}{Size of the leapfrog step}

\item{delta_max}{Measure of the required accuracy of the simulation. The authors recommend a large value (1000)}

\item{iterations}{Number of iterations of the sampler.}
}
\description{
Adapted from Hoffman and Gelman (2014). The No U-Turn Sampler (NUTS) aims to eliminate the need to set a number of steps L that is present in Hamiltonian Monte Carlo, which may lead to undesirable behaviour in HMC if not set correctly.NUTS does so by recursively building a set of candidate points that span the target distribution, and stopping when it starts to double back (hence its name). More information can be found [here](https://arxiv.org/abs/1111.4246)
}
\examples{
NUTS <- sampler_nuts(distr_name = "norm", distr_params = c(0,1), start = 1)
}
