\name{samplr-package}
\alias{samplr-package}
\alias{samplr}
\docType{package}
\title{
\packageTitle{samplr}
}
\description{
\packageDescription{samplr}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{samplr}
\packageIndices{samplr}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{samplr}

Maintainer: \packageMaintainer{samplr}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
