% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplr-package.R
\docType{package}
\name{samplr-package}
\alias{samplr}
\alias{samplr-package}
\title{samplr: Tools To Compare Human Performance To Sampling Algorithms}
\description{
The goal of samplr is to provide tools to understand human performance from the perspective of sampling, both looking at how people generate samples and how people use the samples they have generated. A longer overview and other resources can be found at sampling.warwick.ac.uk.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://lucas-castillo.github.io/samplr}
}

}
\author{
\strong{Maintainer}: Lucas Castillo \email{lucas.castillo@warwick.ac.uk} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
