% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplr-package.R
\docType{package}
\name{samplr-package}
\alias{samplr}
\alias{samplr-package}
\title{samplr: A Minimal Example of Package Creation}
\description{
An example of a minimal package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/paezha/samplr}
  \item Report bugs at \url{https://github.com/paezha/samplr/issues}
}

}
\author{
\strong{Maintainer}: Antonio Paez \email{paezha@mcmaster.ca} (\href{https://orcid.org/0000-0001-6912-9919}{ORCID})

}
\keyword{internal}
