% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplr.R
\name{samplr}
\alias{samplr}
\title{Single and Two Variable Rejection Sampler}
\usage{
samplr(fun, a, b, N, c = NULL, d = NULL)
}
\arguments{
\item{fun}{The pdf that we are sampling from, input as a string. For one variable, the input variable must be declared as `x`. For two variables, the input variables must be declared as `x` and `y`.}

\item{a}{The lower bound of the variable `x` in the pdf.}

\item{b}{The upper bound of the variable `x` in the pdf.}

\item{N}{The number of samples output by the rejection sampler.}

\item{c}{The lower bound of the variable `y` in the pdf.}

\item{d}{The upper bound of the variable `y` in the pdf.}
}
\value{
If using a single variable, the output is a vector containing the samples. If two variables, the output is a matrix with N rows and 2 columns containing the samples.
}
\description{
This function implements both one and two variable rejection sampling for rvs with bounded support and which have bounded pdf.
}
\examples{

Single variable

samps <- samplr("2*x", 0, 1, 10000)
hist(samps)

Two variable

requires(ggplot2)

samps <- samplr("x+y",a = 0,b = 1,c = 0,d = 1,N = 10000)
samps <- data.frame(samps)
colnames(samps) <- c("x","y")
ggplot(samps, aes(x, y)) +
    geom_density_2d()

}
