% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplr-package.R
\docType{package}
\name{samplr}
\alias{samplr-package}
\alias{samplr}
\title{samplr: Spatially balanced sampling}
\description{
Select probability samples in multi-dimensional spaces with any prescribed inclusion
probabilities using fast rust implementations.

You can access the project website at
\url{https://envisim.se}.
}
\references{
Benedetti, R., Piersimoni, F., & Postiglione, P. (2017).
Spatially balanced sampling: a review and a reappraisal.
International Statistical Review, 85(3), 439-454.

Friedman, J. H., Bentley, J. L., & Finkel, R. A. (1977).
An algorithm for finding best matches in logarithmic expected time.
ACM Transactions on Mathematical Software (TOMS), 3(3), 209-226.

Maneewongvatana, S., & Mount, D. M. (1999).
It’s okay to be skinny, if your friends are fat.
In Center for geometric computing 4th annual workshop on computational geometry (Vol. 2, pp. 1-8).

Särndal, C. E., Swensson, B., & Wretman, J. (2003).
Model assisted survey sampling.
Springer Science & Business Media.

Tillé, Y. (2006).
Sampling algorithms.
New York, NY: Springer New York.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://www.envisim.se/}
  \item \url{https://github.com/envisim/rust-samplr/}
  \item Report bugs at \url{https://github.com/envisim/rust-samplr/issues}
}

}
\author{
Wilmer Prentius \email{wilmer.prentius@gmail.com}, Anton Grafström.
}
\keyword{internal}
