% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplr.R
\docType{package}
\name{samplr}
\alias{samplr}
\title{samplr: A package for comparing human performance to sampling algorithms}
\description{
The samplr package provides two categories of functions:
sampler_, plot_.
}
\section{sampler_ functions}{

The sampler_ functions obtain samples from a given distribution following the specified algorithm.
}

\section{plot_ functions}{

The plot_ functions provide diagnostic tools on the performance of either the sampling algorithms, or people.
}

