% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplr.R
\docType{package}
\name{samplr}
\alias{samplr}
\title{mypackage: A package for computating the notorious bar statistic.}
\description{
The mypackage package provides three categories of important functions:
foo, bar and baz.
}
\section{Mypackage functions}{

The mypackage functions ...
}

