% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplr.R
\docType{package}
\name{samplr}
\alias{samplr}
\alias{samplr-package}
\title{samplr: A package for sampling from continuous distributions.}
\description{
The samplr package provides a number of functions for sampling from continuous distributions.
}
