% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABSClass.R
\name{Zhu23ABS}
\alias{Zhu23ABS}
\title{Auto-correlated Bayesian Sampler by Zhu (2023)}
\description{
This Auto-correlated Bayesian Sampler model \insertCite{@ABS, @zhuAutocorrelatedBayesian2023}{samplr} is developed by Zhu.
}
\examples{

## ------------------------------------------------
## Method `Zhu23ABS$new`
## ------------------------------------------------

zhuabs <- Zhu23ABS$new(width = 1, n_chains = 5, nd_time = 0.3, s_nd_time = 0.5, lambda = 10)


## ------------------------------------------------
## Method `Zhu23ABS$simulate`
## ------------------------------------------------


trial_stim <- round(runif(5, 10, 50))
zhuabs$simulate(stopping_rule='fixed', n_sample = 5, trial_stim = trial_stim)
zhuabs$sim_results

zhuabs$reset_sim_results()
trial_stim <- factor(sample(c('left', 'right'), 5, TRUE))
zhuabs$simulate(stopping_rule='relative', delta = 4, dec_bdry = 0, discrim = 1, trial_stim = trial_stim)
zhuabs$sim_results


## ------------------------------------------------
## Method `Zhu23ABS$confidence_interval`
## ------------------------------------------------

zhuabs$confidence_interval(conf_level = 0.9)

}
\references{
\insertRef{zhuAutocorrelatedBayesian2023}{samplr}
}
\section{Super class}{
\code{\link[samplr:CoreABS]{samplr::CoreABS}} -> \code{Zhu23ABS}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{width}}{the standard deviation of the proposal distribution for MC3.}

\item{\code{lambda}}{the rate parameter of the Erlang distribution for decision time.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Zhu23ABS-new}{\code{Zhu23ABS$new()}}
\item \href{#method-Zhu23ABS-simulate}{\code{Zhu23ABS$simulate()}}
\item \href{#method-Zhu23ABS-confidence_interval}{\code{Zhu23ABS$confidence_interval()}}
\item \href{#method-Zhu23ABS-reset_sim_results}{\code{Zhu23ABS$reset_sim_results()}}
\item \href{#method-Zhu23ABS-clone}{\code{Zhu23ABS$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Zhu23ABS-new"></a>}}
\if{latex}{\out{\hypertarget{method-Zhu23ABS-new}{}}}
\subsection{Method \code{new()}}{
Create a new 'Zhu23ABS' object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Zhu23ABS$new(width, n_chains, nd_time, s_nd_time, lambda, distr_name = "norm")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{width}}{the standard deviation of the proposal distribution for MC3.}

\item{\code{n_chains}}{an integer of the number of chains for the sampler.}

\item{\code{nd_time}}{the non-decision time (in seconds).}

\item{\code{s_nd_time}}{the inter-trial-variability of the non-decition time (in seconds).}

\item{\code{lambda}}{the rate parameter of the Erlang distribution for decision time.}

\item{\code{distr_name}}{a character string indicating the type of the posterior hypothesis distribution.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new 'Zhu23ABS' object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{zhuabs <- Zhu23ABS$new(width = 1, n_chains = 5, nd_time = 0.3, s_nd_time = 0.5, lambda = 10)

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Zhu23ABS-simulate"></a>}}
\if{latex}{\out{\hypertarget{method-Zhu23ABS-simulate}{}}}
\subsection{Method \code{simulate()}}{
Simulate the ABS model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Zhu23ABS$simulate(stopping_rule, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{stopping_rule}}{a character string indicating the stopping rule of ABS to be applied. Possible values are \code{"fixed"} and \code{"relative"}. See also \code{Details}.}

\item{\code{...}}{further arguments passed to the ABS model, see also \code{Details}.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
The ABS model has two types of stopping rules: fixed and relative. The fixed stopping rule means that a fixed number of samples are drawn to complete the tasks such as estimations and confidence intervals. On the other hand, the relative stopping rule means that the model counts the difference in evidence between the two hypotheses, and terminates the sampling process whenever the accumulated difference exceeds a threshold. This rule applies to tasks such as two-alternative force choice tasks.

When \verb{stopping rule} is \code{"fixed"}, the following arguments are required:
\itemize{
\item \code{n_sample} the fixed number of samples for each trial.
\item \code{trial_stim} a numeric vector of the stimulus of each trial.
\item \code{start_point} a numeric vector setting the start point of each trial for the sampler. Defaults to \code{NA}, meaning that the start point of each trial is the last sample of the previous trial. Please refer to the vignette for more information.
}

When the stopping rule is "relative", the following arguments are required:
\itemize{
\item \code{delta} the relative difference between the number of samples supporting each hypothesis.
\item \code{dec_bdry} the decision boundary that separates the posterior hypothesis distribution.
\item \code{discrim} the stimuli discriminability.
\item \code{trial_stim} a factor only consisting of two levels: below and above the decision boundary. It indicates the stimulus of each trial.
\item \code{prior_on_resp} a numeric vector for the Beta prior on responses. Defaults to \code{c(1,1)} representing the distribution \code{Beta(1,1)}.
\item \code{start_point} a numeric vector setting the start point of each trial for the sampler. Defaults to \code{NA}, meaning that the start point of each trial is the last sample of the previous trial. Please refer to the vignette for more information.
\item \code{stim_depend} a boolean variable that control whether the prior on responses changes regarding the last stimulus. Defaults to \code{TRUE}. Please refer to the vignette for more information.
\item \code{max_iterations} the maximum length of the MC3 sampler. Defaults to 1000. The program will stop the sampling process after the length of the sampling sequence reaches to this limitation.
}

No values will be return after running this method, but the field \code{sim_results} will be updated instead. If the stopping rule is "fixed", \code{simulation_results} will be a data frame with five columns:
\enumerate{
\item{trial: The index of trials;}
\item{samples: The samples of ABS sampler for the trial;}
\item{stimulus: The stimuli of the experiment;}
\item{rt: The response time;}
\item{point_est: The response of point estimation;}
}

On the other hand, if the stopping rule is "relative", \code{sim_results} will be a data frame with seven columns:
\enumerate{
\item{trial: The index of trials;}
\item{samples: The samples of ABS sampler for the trial;}
\item{response: The response predicted by ABS;}
\item{stimulus: The stimuli of the experiment;}
\item{accuracy: Whether the response is the same as the feedback. 0 represents error, and 1 represents correct;}
\item{rt: The response time, including both the non-decision and the decision time;}
\item{confidence: The confidence of the response.}
}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{
trial_stim <- round(runif(5, 10, 50))
zhuabs$simulate(stopping_rule='fixed', n_sample = 5, trial_stim = trial_stim)
zhuabs$sim_results

zhuabs$reset_sim_results()
trial_stim <- factor(sample(c('left', 'right'), 5, TRUE))
zhuabs$simulate(stopping_rule='relative', delta = 4, dec_bdry = 0, discrim = 1, trial_stim = trial_stim)
zhuabs$sim_results

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Zhu23ABS-confidence_interval"></a>}}
\if{latex}{\out{\hypertarget{method-Zhu23ABS-confidence_interval}{}}}
\subsection{Method \code{confidence_interval()}}{
This function calculates the confidence interval of the \code{simulate} method's results when the "fixed" stopping rule was used.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Zhu23ABS$confidence_interval(conf_level)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{conf_level}}{the required confidence level.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
No values will be returned by this method. Instead, two new columns will be added to the \code{sim_results}:
\enumerate{
\item{conf_interval_l: The lower bound of the confidence interval with the given level;}
\item{conf_interval_u: The upper bound of the confidence interval with the given level;}
}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{zhuabs$confidence_interval(conf_level = 0.9)

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Zhu23ABS-reset_sim_results"></a>}}
\if{latex}{\out{\hypertarget{method-Zhu23ABS-reset_sim_results}{}}}
\subsection{Method \code{reset_sim_results()}}{
This function is for resetting the \code{ss_samples} to run new simulations.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Zhu23ABS$reset_sim_results()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Zhu23ABS-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Zhu23ABS-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Zhu23ABS$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
