% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABSClass.R
\name{Zhu23ABS}
\alias{Zhu23ABS}
\title{Auto-correlated Bayesian Sampler by Zhu (2023)}
\description{
This Auto-correlated Bayesian Sampler model \insertCite{@ABS, @zhuAutocorrelatedBayesian2023}{samplr} is developed by Zhu.
}
\examples{

## ------------------------------------------------
## Method `Zhu23ABS$new`
## ------------------------------------------------

zhuabs <- Zhu23ABS$new(nd_time = 0.3, s_nd_time = 0.5, lambda = 10, n_chains = 5, width = 1, distr_name = 'norm')


## ------------------------------------------------
## Method `Zhu23ABS$two_alt_force_choice`
## ------------------------------------------------

trial_stim <- factor(sample(c('left', 'right'), 5, TRUE))
tafc_sim <- zhuabs$two_alt_force_choice(delta = 4, dec_bdry = 0, discrim = 1, trial_stim = trial_stim)


## ------------------------------------------------
## Method `Zhu23ABS$estimate`
## ------------------------------------------------

trial_stim <- c(25, 26, 10, 30)
est_sim <- zhuabs$estimate(n_sample = 5, trial_stim = trial_stim, conf_level = 0.5)

}
\references{
\insertAllCited{}
}
\section{Super class}{
\code{\link[samplr:CoreABS]{samplr::CoreABS}} -> \code{Zhu23ABS}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{lambda}}{The rate parameter of the gamma distribution for decision time.}

\item{\code{width}}{The standard deviation of the proposal distribution for MC3}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Zhu23ABS-new}{\code{Zhu23ABS$new()}}
\item \href{#method-Zhu23ABS-two_alt_force_choice}{\code{Zhu23ABS$two_alt_force_choice()}}
\item \href{#method-Zhu23ABS-estimate}{\code{Zhu23ABS$estimate()}}
\item \href{#method-Zhu23ABS-clone}{\code{Zhu23ABS$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Zhu23ABS-new"></a>}}
\if{latex}{\out{\hypertarget{method-Zhu23ABS-new}{}}}
\subsection{Method \code{new()}}{
Create a new 'Zhu23ABS' object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Zhu23ABS$new(
  nd_time,
  s_nd_time = 0,
  lambda,
  n_chains,
  width,
  distr_name = "norm",
  start_point = NA
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{nd_time}}{The non-decision time.}

\item{\code{s_nd_time}}{The range of the non-decision time. Default is 0, implying a fixed non-decision time.}

\item{\code{lambda}}{The rate parameter of the gamma distribution for decision time.}

\item{\code{n_chains}}{The number of chains of the sampler. It should be an integer.}

\item{\code{width}}{The proposal width of the sampler.}

\item{\code{distr_name}}{The type of the posterior hypothesis distribution.}

\item{\code{start_point}}{The start point of each trial.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new 'Zhu23ABS' object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{zhuabs <- Zhu23ABS$new(nd_time = 0.3, s_nd_time = 0.5, lambda = 10, n_chains = 5, width = 1, distr_name = 'norm')

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Zhu23ABS-two_alt_force_choice"></a>}}
\if{latex}{\out{\hypertarget{method-Zhu23ABS-two_alt_force_choice}{}}}
\subsection{Method \code{two_alt_force_choice()}}{
This function is for simulating two-alternative-force choice tasks by Zhu23ABS.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Zhu23ABS$two_alt_force_choice(
  delta,
  dec_bdry,
  prior_on_resp = c(1, 1),
  discrim,
  trial_stim,
  stim_depend = TRUE,
  max_iterations = 1000
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{delta}}{The relative difference between the number of samples supporting each hypothesis.}

\item{\code{dec_bdry}}{The decision boundary that separates the posterior hypothesis distribution.}

\item{\code{prior_on_resp}}{The beta prior on responses. Default setting is Beta(1,1)}

\item{\code{discrim}}{The stimuli discriminability.}

\item{\code{trial_stim}}{The stimulus of each trial. It should be a factor only consisting of two levels: below and above the decision boundary.}

\item{\code{stim_depend}}{The boolean variable that control whether the prior on responses changes regarding the last stimulus.}

\item{\code{max_iterations}}{The maximum length of the MCREC sampler. The program will stop the sampling process after the length of the sampling sequence reaches to this limitation.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data frame with seven columns:
\enumerate{
\item{trial: The index of trials;}
\item{samples: The samples of ABS sampler for the trial;}
\item{response: The response predicted by ABS;}
\item{stimulus: The stimuli of the experiment;}
\item{accuracy: Whether the response is the same as the feedback. 0 represents error, and 1 represents correct;}
\item{rt: The response time, including both the non-decision and the decision time;}
\item{confidence: The confidence of the response.}
}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{trial_stim <- factor(sample(c('left', 'right'), 5, TRUE))
tafc_sim <- zhuabs$two_alt_force_choice(delta = 4, dec_bdry = 0, discrim = 1, trial_stim = trial_stim)

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Zhu23ABS-estimate"></a>}}
\if{latex}{\out{\hypertarget{method-Zhu23ABS-estimate}{}}}
\subsection{Method \code{estimate()}}{
This function is for simulating the estimate and confidence interval tasks by Zhu23ABS.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Zhu23ABS$estimate(n_sample, trial_stim, conf_level = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n_sample}}{The fixed number of samples for each trial.}

\item{\code{trial_stim}}{The stimulus of each trial.}

\item{\code{conf_level}}{The required confidence level.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data frame with ten columns:
\enumerate{
\item{trial: The index of trials;}
\item{samples: The samples of ABS sampler for the trial;}
\item{stimulus: The stimuli of the experiment;}
\item{nd_time: The non-decision time for this trial;}
\item{conf_interval_l: The lower bound of the confidence interval with the given level;}
\item{conf_interval_u: The upper bound of the confidence interval with the given level;}
}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{trial_stim <- c(25, 26, 10, 30)
est_sim <- zhuabs$estimate(n_sample = 5, trial_stim = trial_stim, conf_level = 0.5)

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Zhu23ABS-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Zhu23ABS-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Zhu23ABS$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
