% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABSClass.R
\name{Zhu23ABS}
\alias{Zhu23ABS}
\title{Auto-correlated Bayesian Sampler by Zhu (2023)}
\description{
This Auto-correlated Bayesian Sampler model \insertCite{@ABS, @zhuAutocorrelatedBayesian2023}{samplr} is developed by Zhu.
}
\examples{

## ------------------------------------------------
## Method `Zhu23ABS$new`
## ------------------------------------------------

abs_model <- Zhu23ABS$new(nd_time=0.3, s_nd_time=0.5, prior_on_resp=c(1, 1), lambda=10, n_chains=5, width=1, distr_name='norm')


## ------------------------------------------------
## Method `Zhu23ABS$two_alt_force_choice`
## ------------------------------------------------

trial_stim <- factor(c('left', 'right', 'left', 'left', 'right'))
start_point <- as.matrix(stats::runif(5, min=-3, max=3))
tafc_sim <- abs_model$two_alt_force_choice(delta=4, dec_bdry=0, discrim=1, trial_stim=trial_stim, start_point=start_point)


## ------------------------------------------------
## Method `Zhu23ABS$estimate`
## ------------------------------------------------

start_point <- as.matrix(abs(stats::rnorm(5, 0, 100)))
est_sim <- abs_model$estimate(n_sample = 5, trial_stim = c(25, 26, 25, 24), start_point, conf_level=0.5)

}
\references{
\insertRef{zhuAutocorrelatedBayesian2023}{samplr}
}
\section{Super class}{
\code{\link[samplr:CoreABS]{samplr::CoreABS}} -> \code{Zhu23ABS}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{width}}{The standard deviation of the proposal distribution for MC3}

\item{\code{nd_time}}{The non-decision time.}

\item{\code{s_nd_time}}{The range of the non-decision time.}

\item{\code{prior_on_resp}}{The beta prior on responses.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Zhu23ABS-new}{\code{Zhu23ABS$new()}}
\item \href{#method-Zhu23ABS-two_alt_force_choice}{\code{Zhu23ABS$two_alt_force_choice()}}
\item \href{#method-Zhu23ABS-estimate}{\code{Zhu23ABS$estimate()}}
\item \href{#method-Zhu23ABS-clone}{\code{Zhu23ABS$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Zhu23ABS-new"></a>}}
\if{latex}{\out{\hypertarget{method-Zhu23ABS-new}{}}}
\subsection{Method \code{new()}}{
Create a new 'Zhu23ABS' object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Zhu23ABS$new(
  nd_time,
  s_nd_time = 0,
  prior_on_resp = c(1, 1),
  lambda,
  n_chains,
  width,
  distr_name = "norm"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{nd_time}}{The non-decision time.}

\item{\code{s_nd_time}}{The range of the non-decision time. Default is 0, implying a fixed non-decision time.}

\item{\code{prior_on_resp}}{The beta prior on responses. Default setting is Beta(1,1)}

\item{\code{lambda}}{The rate parameter of the gamma distribution for decision time.}

\item{\code{n_chains}}{The number of chains of the sampler. It should be an integer.}

\item{\code{width}}{The proposal width of the sampler.}

\item{\code{distr_name}}{The type of the posterior hypothesis distribution.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new 'Zhu23ABS' object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{abs_model <- Zhu23ABS$new(nd_time=0.3, s_nd_time=0.5, prior_on_resp=c(1, 1), lambda=10, n_chains=5, width=1, distr_name='norm')

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Zhu23ABS-two_alt_force_choice"></a>}}
\if{latex}{\out{\hypertarget{method-Zhu23ABS-two_alt_force_choice}{}}}
\subsection{Method \code{two_alt_force_choice()}}{
This function is for simulating two-alternative-force choice tasks by Zhu23ABS.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Zhu23ABS$two_alt_force_choice(
  delta,
  dec_bdry,
  discrim,
  trial_stim,
  start_point,
  mc3_iterations = 100
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{delta}}{The relative difference between the number of samples supporting each hypothesis.}

\item{\code{dec_bdry}}{The decision boundary that separates the posterior hypothesis distribution.}

\item{\code{discrim}}{The stimuli discriminability.}

\item{\code{trial_stim}}{The stimulus of each trial. It should be a factor only consisting of two levels: below and above the decision boundary.}

\item{\code{start_point}}{The start point of the sampler.}

\item{\code{mc3_iterations}}{The number of iterations for each chunk. See details for more information.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
The \code{two_alt_force_choice} function in the Zhu23ABS class runs simulations in chunks of fixed length determined by the \code{mc3_iterations} argument. In each trial simulation, the function generates the first chunk and checks if the simulated sequence meets the stopping rule within that chunk. If the stopping rule is met, the function cuts the chunk at that point, returns the sequence, and starts a new trial simulation. If the stopping rule is not met, the function proceeds to the next chunk, beginning where the previous chunk ended, and repeats the entire process.
}

\subsection{Returns}{
A data frame with seven columns:
\enumerate{
\item{trial: The index of trials;}
\item{samples: The samples of ABS sampler for the trial;}
\item{response: The response predicted by ABS;}
\item{stimulus: The stimuli of the experiment;}
\item{accuracy: Whether the response is the same as the feedback. 0 represents error, and 1 represents correct;}
\item{rt: The response time, including both the non-decision and the decision time;}
\item{confidence: The confidence of the response.}
}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{trial_stim <- factor(c('left', 'right', 'left', 'left', 'right'))
start_point <- as.matrix(stats::runif(5, min=-3, max=3))
tafc_sim <- abs_model$two_alt_force_choice(delta=4, dec_bdry=0, discrim=1, trial_stim=trial_stim, start_point=start_point)

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Zhu23ABS-estimate"></a>}}
\if{latex}{\out{\hypertarget{method-Zhu23ABS-estimate}{}}}
\subsection{Method \code{estimate()}}{
This function is for simulating the estimate and confidence interval tasks by Zhu23ABS.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Zhu23ABS$estimate(n_sample, trial_stim, start_point, conf_level = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n_sample}}{The fixed number of samples for each trial.}

\item{\code{trial_stim}}{The stimulus of each trial.}

\item{\code{start_point}}{The start point of the sampler.}

\item{\code{conf_level}}{The required confidence level.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data frame with ten columns:
\enumerate{
\item{trial: The index of trials;}
\item{samples: The samples of ABS sampler for the trial;}
\item{stimulus: The stimuli of the experiment;}
\item{nd_time: The non-decision time for this trial;}
\item{conf_interval_l: The lower bound of the confidence interval with the given level;}
\item{conf_interval_u: The upper bound of the confidence interval with the given level;}
}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{start_point <- as.matrix(abs(stats::rnorm(5, 0, 100)))
est_sim <- abs_model$estimate(n_sample = 5, trial_stim = c(25, 26, 25, 24), start_point, conf_level=0.5)

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Zhu23ABS-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Zhu23ABS-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Zhu23ABS$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
